pt_int struct now uses long long as internal saving.

+++ src/pillowtalk.h
@@ -64,7 +64,7 @@ pt_node_t* pt_array_get(pt_node_t* array, unsigned int idx);
 
 int pt_is_null(pt_node_t* null);
 int pt_boolean_get(pt_node_t* boolean);
-int pt_integer_get(pt_node_t* integer);
+long long pt_integer_get(pt_node_t* integer);
 double pt_double_get(pt_node_t* dbl);
 const char* pt_string_get(pt_node_t* string);
